Added cors configs and improved authorization

+++ auth-service/src/main/java/com/aonufrei/auth_service/rest/AuthRestController.java
@@ -1,21 +1,21 @@
 package com.aonufrei.auth_service.rest;
 
 import com.aonufrei.auth_service.service.AccountService;
-import com.aonufrei.dto.AccountOutDto;
-import com.aonufrei.dto.ChangePasswordRequestDto;
-import com.aonufrei.dto.LoginDto;
-import com.aonufrei.dto.RegisterDto;
+import com.aonufrei.dto.*;
 import com.aonufrei.enums.AccountRole;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.http.HttpStatus;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Objects;
 
 @RestController
 @RequestMapping("api/v1/auth")
 public class AuthRestController {
 
+	private final Logger log = LoggerFactory.getLogger(AuthRestController.class);
 	private final AccountService accountService;
 
 	public AuthRestController(AccountService accountService) {
@@ -23,8 +23,8 @@ public class AuthRestController {
 	}
 
 	@PostMapping("login")
-	public String login(@RequestParam AccountRole role, @Validated @RequestBody LoginDto loginDto) {
-		return accountService.login(role, loginDto);
+	public LoginOutDto login(@Validated @RequestBody LoginDto loginDto) {
+		return accountService.login(loginDto);
 	}
 
 	@PostMapping("register")
@@ -32,27 +32,29 @@ public class AuthRestController {
 		return accountService.register(role, registerDto);
 	}
 
-	@GetMapping("identify")
-	public ResponseEntity<AccountOutDto> identifyUser(HttpServletRequest servletRequest) {
-		String token = servletRequest.getHeader("Authorization");
-		if (token == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
-		return ResponseEntity.ok(accountService.identifyUser(token));
-	}
-
 	@PutMapping("change-password")
 	public ResponseEntity<Boolean> changePassword(@Validated @RequestBody ChangePasswordRequestDto payload) {
+		var auth = AccountService.getCurrentAuth();
+		if (Objects.equals(payload.getAccountId(), auth.id())) {
+			return ResponseEntity.status(404).body(false);
+		}
 		accountService.changePassword(payload);
 		return ResponseEntity.ok(true);
 	}
 
 	@DeleteMapping
 	public ResponseEntity<Boolean> deleteAccount(@RequestParam String accountId) {
+		var auth = AccountService.getCurrentAuth();
+		if (Objects.equals(accountId, auth.id())) {
+			return ResponseEntity.status(404).body(false);
+		}
 		accountService.deleteAccount(accountId);
 		return ResponseEntity.ok(true);
 	}
 
 	@ExceptionHandler(RuntimeException.class)
 	public ResponseEntity<String> validationErrorHandler(RuntimeException e) {
+		log.error("An error occurred while processing request", e);
 		return ResponseEntity.badRequest().body(e.getMessage());
 	}
